+++ R/ternary_contour.R
@@ -10,13 +10,16 @@ setMethod(
   signature = c(x = "numeric", y = "numeric", z = "numeric"),
   definition = function(x, y, z, value, n = 50, nlevels = 10,
                         levels = pretty(range(value, na.rm = TRUE), nlevels),
-                        ilr = TRUE, linear = TRUE, extrapolate = FALSE,
+                        ilr = TRUE, method = "linear", extrapolate = FALSE,
                         palette = function(i) grDevices::hcl.colors(i, "YlOrRd", rev = TRUE),
                         ...) {
+    ## /!\ For backward compatibility with akima::interp() /!\
+    if (isTRUE(list(...)$linear)) method <- "linear"
+
     ## Calculate contour lines
     xy <- coordinates_contour(x = x, y = y, z = z, value = value, n = n,
                               nlevels = nlevels, levels = levels,
-                              ilr = ilr, linear = linear,
+                              ilr = ilr, method = method,
                               extrapolate = extrapolate)
 
     ## Get contour levels
@@ -54,13 +57,13 @@ setMethod(
   signature = c(x = "ANY", y = "missing", z = "missing"),
   definition = function(x, value, n = 50, nlevels = 10,
                         levels = pretty(range(value, na.rm = TRUE), nlevels),
-                        ilr = TRUE, linear = TRUE, extrapolate = FALSE,
+                        ilr = TRUE, method = "linear", extrapolate = FALSE,
                         palette = function(i) grDevices::hcl.colors(i, "YlOrRd", rev = TRUE),
                         ...) {
     x <- grDevices::xyz.coords(x)
     methods::callGeneric(x = x$x, y = x$y, z = x$z, value = value,
                          n = n, nlevels = nlevels, levels = levels,
-                         ilr = ilr, linear = linear, extrapolate = extrapolate,
+                         ilr = ilr, method = method, extrapolate = extrapolate,
                          palette = palette, ...)
   }
 )
@@ -78,11 +81,11 @@ setMethod(
 #' @param levels A [`numeric`] vector of levels at which to draw contour lines.
 #' @param ilr A [`logical`] scalar: should interpolation be computed in ILR
 #'  space? If `FALSE`, interpolation is computed in Cartesian space.
-#' @param linear A [`logical`] scalar: should linear interpolation be used?
-#'  If `FALSE`, spline interpolation is used (see [akima::interp()]).
+#' @param method A [`character`] string: specifying the method for interpolation
+#'  (see [interp::interp()]).
 #' @param extrapolate A [`logical`] scalar: should extrapolation be used outside
-#'  of the convex hull determined by the data points (see [akima::interp()])?
-#' @param ... Further parameters to be passed to [akima::interp()].
+#'  of the convex hull determined by the data points (see [interp::interp()])?
+#' @param ... Further parameters to be passed to [interp::interp()].
 #' @return
 #'  A [`list`] of contours, each itself a list with elements
 #'  (see [grDevices::contourLines()]):
@@ -95,18 +98,14 @@ setMethod(
 #' @noRd
 coordinates_contour <- function(x, y, z, value, n = 50, nlevels = 10,
                                 levels = pretty(range(value, na.rm = TRUE), nlevels),
-                                ilr = TRUE, linear = TRUE, extrapolate = FALSE,
+                                ilr = TRUE, method = "linear", extrapolate = FALSE,
                                 ...) {
   ## Validation
-  if (!requireNamespace("akima", quietly = TRUE)) {
-    msg <- "The akima package is required. Please install it."
+  if (!requireNamespace("interp", quietly = TRUE)) {
+    msg <- "The interp package is required. Please install it."
     stop(msg, call. = FALSE)
   }
-  if (length(value) != length(x)) {
-    msg <- sprintf("%s must be of length %d; not %s.",
-                   sQuote("value"), length(x), length(value))
-    stop(msg, call. = FALSE)
-  }
+  assert_length(value, length(x))
 
   ## ILR vs Cartesian
   coda <- cbind(x, y, z)
@@ -120,13 +119,13 @@ coordinates_contour <- function(x, y, z, value, n = 50, nlevels = 10,
   ## Interpolate
   xlim <- expand_range(ratio[, 1], mult = 0.2)
   ylim <- expand_range(ratio[, 2], mult = 0.2)
-  interp <- akima::interp(
+  interp <- interp::interp(
     x = ratio[, 1],
     y = ratio[, 2],
     z = value,
     xo = seq(xlim[1L], xlim[2L], length.out = n),
     yo = seq(ylim[1L], ylim[2L], length.out = n),
-    linear = linear,
+    method = method,
     extrap = extrapolate,
     ...
   )